Drop redundant i64 cast

+++ src/ir/lower.rs
@@ -24334,7 +24334,7 @@ fn lower_array_expr_descriptor(
             };
             if derived_type.is_some() {
                 let zero32 = b.const_i32(0);
-                let total_bytes = b.const_i64(ir_scalar_byte_size(&elem_ty) * n.max(1) as i64);
+                let total_bytes = b.const_i64(ir_scalar_byte_size(&elem_ty) * n.max(1));
                 b.call(
                     FuncRef::External("memset".into()),
                     vec![buf, zero32, total_bytes],
